+++ ios/InteractiveReader/InteractiveReader/Features/Playback/JobPlaybackView.swift
@@ -662,10 +662,8 @@ struct JobPlaybackView: View {
         if shouldIgnoreStaleAppleMusicPauseAfterReaderPlay {
             return false
         }
-        if musicOwnership.ownershipState == .appleMusicBed &&
-            !musicOwnership.isPlaying &&
-            !musicOwnership.isSystemPlaybackPlaying {
-            return true
+        if musicOwnership.ownershipState == .appleMusicBed {
+            return false
         }
         #endif
         #if os(tvOS)