+++ ios/InteractiveReader/InteractiveReader/Features/Playback/LibraryPlaybackView.swift
@@ -652,10 +652,8 @@ struct LibraryPlaybackView: View {
         if shouldIgnoreStaleAppleMusicPauseAfterReaderPlay {
             return false
         }
-        if musicOwnership.ownershipState == .appleMusicBed &&
-            !musicOwnership.isPlaying &&
-            !musicOwnership.isSystemPlaybackPlaying {
-            return true
+        if musicOwnership.ownershipState == .appleMusicBed {
+            return false
         }
         #endif
         #if os(tvOS)